Move attribute utils to attribute module
