+++ saleor/graphql/product/tests/test_product_pagination.py
@@ -4,6 +4,7 @@ from decimal import Decimal
 import graphene
 import pytest
 
+from ....attribute.utils import associate_attribute_values_to_instance
 from ....product.models import (
     Category,
     Collection,
@@ -11,7 +12,6 @@ from ....product.models import (
     ProductType,
     ProductVariant,
 )
-from ....product.utils.attributes import associate_attribute_values_to_instance
 from ....warehouse.models import Stock
 from ...tests.utils import get_graphql_content
