+++ saleor/product/tests/test_generate_name_for_variant.py
@@ -5,12 +5,10 @@ import pytest
 
 from ...attribute import AttributeInputType
 from ...attribute.models import AttributeValue
+from ...attribute.utils import associate_attribute_values_to_instance
 from ..models import Product, ProductType, ProductVariant
 from ..tasks import _update_variants_names
-from ..utils.attributes import (
-    associate_attribute_values_to_instance,
-    generate_name_for_variant,
-)
+from ..utils.variants import generate_name_for_variant
 
 
 @pytest.fixture()
@@ -94,46 +92,3 @@ def test_update_variants_changed_does_nothing_with_no_attributes():
     saved_attributes = []
     # FIXME: This method no longer returns any value
     assert _update_variants_names(product_type, saved_attributes) is None
-
-
-def test_associate_attribute_to_non_product_instance(color_attribute):
-    instance = ProductType()
-    attribute = color_attribute
-    value = color_attribute.values.first()
-
-    with pytest.raises(AssertionError) as exc:
-        associate_attribute_values_to_instance(instance, attribute, value)  # noqa
-
-    assert exc.value.args == ("ProductType is unsupported",)
-
-
-def test_associate_attribute_to_product_instance_from_different_attribute(
-    product, color_attribute, size_attribute
-):
-    """Ensure an assertion error is raised when one tries to associate attribute values
-    to an object that don't belong to the supplied attribute.
-    """
-    instance = product
-    attribute = color_attribute
-    value = size_attribute.values.first()
-
-    with pytest.raises(AssertionError) as exc:
-        associate_attribute_values_to_instance(instance, attribute, value)
-
-    assert exc.value.args == ("Some values are not from the provided attribute.",)
-
-
-def test_associate_attribute_to_product_instance_without_values(product):
-    """Ensure clearing the values from a product is properly working."""
-    old_assignment = product.attributes.first()
-    assert old_assignment is not None, "The product doesn't have attribute-values"
-    assert old_assignment.values.count() == 1
-
-    attribute = old_assignment.attribute
-
-    # Clear the values
-    new_assignment = associate_attribute_values_to_instance(product, attribute)
-
-    # Ensure the values were cleared and no new assignment entry was created
-    assert new_assignment.pk == old_assignment.pk
-    assert new_assignment.values.count() == 0
