+++ saleor/product/tests/test_product.py
@@ -7,11 +7,11 @@ from freezegun import freeze_time
 from prices import Money
 
 from ...account import events as account_events
+from ...attribute.utils import associate_attribute_values_to_instance
 from .. import models
 from ..filters import filter_products_by_attributes_values
 from ..models import DigitalContentUrl
 from ..thumbnails import create_product_thumbnails
-from ..utils.attributes import associate_attribute_values_to_instance
 from ..utils.costs import get_margin_for_variant
 from ..utils.digital_products import increment_download_count
